backend/requirements.txt
sounddevice>=0.4.6
scipy>=1.10.0
numpy>=1.24.0
pynput>=1.7.6
groq>=0.4.2
PySide6>=6.5.0
pyperclip>=1.8.2
python-dotenv>=1.2.1
fastapi>=0.109.0
uvicorn>=0.27.0
websockets>=13.0
pydantic>=2.0.0
ten-vad>=1.0.6
pyobjc-framework-Quartz>=10.0; sys_platform == "darwin"
